requirements.txt
# UltraQuality Screener - Core Dependencies
# Lightweight install for production use

pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
requests>=2.31.0,<3.0.0
pyyaml>=6.0,<7.0
scipy>=1.11.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0

# Testing (optional - comment out for production)
# pytest>=7.4.0,<9.0.0
